from u64 to u32|

+++ src/entities.rs
@@ -8,7 +8,7 @@ use std::fmt::Formatter;
 pub struct CommandPacket {
     pub protocol_id: u8,
     pub protocol_version: u8,
-    pub  id: u64,
+    pub  id: u32,
     pub  command: Vec<u8>,
 }
 
@@ -22,8 +22,8 @@ impl CommandPacket {
 pub struct StatePacket {
     pub protocol_id: u8,
     pub protocol_version: u8,
-    pub  id: u64,
-    pub lost_ids: Vec<u64>,
+    pub  id: u32,
+    pub lost_ids: Vec<u32>,
     pub state: Vec<u8>,
 }
